+++ pkg/cloudstorage/main_test.go
