+++ pkg/cloudstorage/path.go
