+++ pkg/cloudstorage/path_test.go
